Refactor popup and GettingStart components; remove unused files and styles for cleaner codebase
