--- app/css/components/GettingStart.css
@@ -1,36 +1,0 @@
-.getting-start-block {
-  padding: 40px 0 32px 0;
-  text-align: center;
-  color: #666;
-  background: #f8fafc;
-  border-radius: 12px;
-  margin: 32px 0 24px 0;
-  box-shadow: 0 2px 12px 0 rgba(0,0,0,0.04);
-  border: 1px solid #e3e8ee;
-  max-width: 320px;
-  margin-left: auto;
-  margin-right: auto;
-}
-.getting-start-block .getting-start-title {
-  margin-bottom: 18px;
-  font-size: 1.15em;
-  font-weight: 500;
-  letter-spacing: 0.01em;
-}
-.getting-start-block .getting-start-title .icon {
-  display: inline-block;
-  vertical-align: middle;
-  margin-right: 8px;
-  font-size: 1.5em;
-}
-.getting-start-block .getting-start-desc {
-  margin-bottom: 22px;
-  color: #888;
-  font-size: 0.98em;
-}
-.getting-start-block .btn {
-  padding: 8px 22px;
-  font-size: 1em;
-  border-radius: 6px;
-  box-shadow: 0 1px 4px 0 rgba(0,0,0,0.06);
-}
