+++ app/manifest.json
@@ -11,7 +11,7 @@
     "128": "images/icon-128.png"
   },
   "background": {
-    "service_worker": "scripts/background.js"
+    "service_worker": "background.js"
   },
   "action": {
     "default_icon": {
@@ -19,7 +19,7 @@
       "48": "images/icon-48.png"
     },
     "default_title": "__MSG_browserActionTitle__",
-    "default_popup": "pages/popup.html"
+    "default_popup": "popup.html"
   },
   "permissions": [
     "activeTab"
