+++ app/popup.css
@@ -1,0 +1,2 @@
+@import url('../components/PopupApp.css');
+@import url('../components/GettingStart.css');
