+++ app/popup.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="utf-8" />
     <title>Popup</title>
-    <link href="../styles/popup.css" rel="stylesheet" type="text/css" />
-    <script src="../scripts/popup.js" type="module"></script>
+    <link href="popup.css" rel="stylesheet" type="text/css" />
+    <script src="popup.js" type="module"></script>
   </head>
   <body>
     <div id="app"></div>
